исправлены ошибки

+++ src/main/java/com/kuxoca/mironline/repo/CurrencyRepo.java
@@ -5,6 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface CurrencyRepo extends JpaRepository<Currency, Long> {
@@ -13,6 +14,6 @@ public interface CurrencyRepo extends JpaRepository<Currency, Long> {
     List<String> findDistinctByName();
 
     @Query(value = "select g.currency from Currency g where (g.localDateTime = (select max(u.localDateTime) from Currency u where u.name = :name) and g.name = :name)")
-    Float findLastDataByName(@Param("name") String name);
+    BigDecimal findLastDataByName(@Param("name") String name);
 
 }
